requirements.txt
flask
groq
pytesseract
pillow
numpy